requirement_spa.txt
absl-py==2.3.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.1
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
autopage==0.5.2
black==25.9.0
certifi==2025.10.5
cfgv==3.4.0
click==8.3.0
cliff==4.11.0
cmaes==0.12.0
cmd2==2.7.0
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
einops==0.8.0
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.11.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
identify==2.6.15
idna==3.11
imageio==2.34.1
importlib_metadata==8.7.0
iniconfig==2.3.0
isort==5.13.2
Jinja2==3.1.6
kiwisolver==1.4.9
lightning==2.5.5
lightning-utilities==0.15.2
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.5
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
opencv-python==4.10.0.82
optuna==2.10.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
pre-commit==3.7.1
prettytable==3.16.0
propcache==0.4.1
protobuf==6.33.0
Pygments==2.19.2
pyparsing==3.2.5
pyperclip==1.11.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytokens==0.2.0
pytorch-lightning==2.5.5
PyYAML==6.0.1
rich==13.7.1
rich-argparse==1.7.2
rootutils==1.0.7
safetensors==0.6.2
scipy==1.13.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
-e git+https://github.com/HaoyiZhu/SPA.git@37b7e0477d7deb031f59a647adffb7b0e2f761f4#egg=spa
SQLAlchemy==2.0.44
stevedore==5.5.0
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
timm==0.9.10
tomli==2.3.0
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchmetrics==1.8.2
torchvision==0.17.1+cu118
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
virtualenv==20.35.4
wcwidth==0.2.14
Werkzeug==3.1.3
yapf==0.40.2
yarl==1.22.0
zipp==3.23.0
